Make the generators objects internal and part of the Generators namespace
