+++ Sources/Genything/GeneratorSequence.swift
