+++ Sources/Genything/Generators/Compositional/Generators+either.swift
@@ -15,32 +15,33 @@ extension Generators {
         where G1: Generator, G2: Generator, G1.T == G2.T {
         Either(left, right, probability: rightProbability).eraseToAnyGenerator()
     }
-}
 
-// MARK: - Either
+    // MARK: - Either
 
-private struct Either<A, B>: Generator where A: Generator, B: Generator, A.T == B.T {
+    struct Either<A, B>: Generator where A: Generator, B: Generator, A.T == B.T {
 
-    public init(_ a: A, _ b: B, probability: Double = 0.5) {
-        self.a = a
-        self.b = b
-        self.probability = probability
-    }
-
-    public let a: A
-    public let b: B
-    public let probability: Double
-
-    public func next(_ randomSource: RandomSource) -> A.T {
-        let value = Double.random(in: probabilityRange, using: &randomSource.rng)
-
-        if value <= probability {
-            return b.next(randomSource)
+        public init(_ a: A, _ b: B, probability: Double = 0.5) {
+            self.a = a
+            self.b = b
+            self.probability = probability
         }
 
-        return a.next(randomSource)
-    }
+        public let a: A
+        public let b: B
+        public let probability: Double
 
-    private let probabilityRange: ClosedRange<Double> = 0.0 ... 1.0
+        public func next(_ randomSource: RandomSource) -> A.T {
+            let value = Double.random(in: probabilityRange, using: &randomSource.rng)
+
+            if value <= probability {
+                return b.next(randomSource)
+            }
+
+            return a.next(randomSource)
+        }
+
+        private let probabilityRange: ClosedRange<Double> = 0.0 ... 1.0
+
+    }
 
 }
