+++ Sources/Genything/Generators/Compositional/Generators+reduce.swift
