+++ Sources/Genything/Generators/Data/Generators+loop.swift
@@ -16,39 +16,40 @@ extension Generators {
     public static func loop<T>(_ collection: T...) -> AnyGenerator<T> {
         loop(collection)
     }
-}
 
-/// A stateful generator which produces elements by looping over a collection of elements.
-///
-/// Since this generator will comprehensibly examine all elements of the collection it can be used to display all possible configurations or test all possible values when the problem space is small and known.
-///
-/// - Warning: This generator is stateful and cannot be restarted. Pay attention to how the reference is being retained and do not share the generator to unsuspecting consumers.
-class Loop<Elements>: Generator where Elements: Collection {
 
-    /// Creates a Generator that loops the provided collection of elements.
+    /// A stateful generator which produces elements by looping over a collection of elements.
     ///
-    /// - Precondition: The collection must not be empty.
-    /// - Parameter collection: The collection of elements to loop.
-    public init(_ collection: Elements) {
-        assert(collection.startIndex != collection.endIndex)
+    /// Since this generator will comprehensibly examine all elements of the collection it can be used to display all possible configurations or test all possible values when the problem space is small and known.
+    ///
+    /// - Warning: This generator is stateful and cannot be restarted. Pay attention to how the reference is being retained and do not share the generator to unsuspecting consumers.
+    class Loop<Elements>: Generator where Elements: Collection {
 
-        self.collection = collection
-        index = collection.startIndex
-    }
+        /// Creates a Generator that loops the provided collection of elements.
+        ///
+        /// - Precondition: The collection must not be empty.
+        /// - Parameter collection: The collection of elements to loop.
+        public init(_ collection: Elements) {
+            assert(collection.startIndex != collection.endIndex)
 
-    /// The collection from which this generator takes it's values
-    public let collection: Elements
-    /// An index which tracks our position in the collection
-    public private(set) var index: Elements.Index
-
-    /// Produces the next element from this generator
-    public func next(_: RandomSource) -> Elements.Element {
-        defer {
-            collection.formIndex(after: &index)
-        }
-        if index >= collection.endIndex {
+            self.collection = collection
             index = collection.startIndex
         }
-        return collection[index]
+
+        /// The collection from which this generator takes it's values
+        public let collection: Elements
+        /// An index which tracks our position in the collection
+        public private(set) var index: Elements.Index
+
+        /// Produces the next element from this generator
+        public func next(_: RandomSource) -> Elements.Element {
+            defer {
+                collection.formIndex(after: &index)
+            }
+            if index >= collection.endIndex {
+                index = collection.startIndex
+            }
+            return collection[index]
+        }
     }
 }
