+++ Sources/Genything/Operators/Generator+flatMap.swift
@@ -8,15 +8,16 @@ extension Generator {
     public func flatMap<G: Generator>(_ transform: @escaping (T) -> G) -> AnyGenerator<G.T> {
         FlatMap(source: self, transform: transform).eraseToAnyGenerator()
     }
-}
 
-// MARK: - FlatMap
 
-private struct FlatMap<Source, Output>: Generator where Source: Generator, Output: Generator {
-    let source: Source
-    let transform: (Source.T) -> Output
 
-    public func next(_ randomSource: RandomSource) -> Output.T {
-        transform(source.next(randomSource)).next(randomSource)
+    struct FlatMap<Source, Output>: Generator where Source: Generator, Output: Generator {
+        let source: Source
+        let transform: (Source.T) -> Output
+
+        public func next(_ randomSource: RandomSource) -> Output.T {
+            transform(source.next(randomSource)).next(randomSource)
+        }
     }
+
 }
