+++ Sources/Genything/Operators/Generator+map.swift
@@ -8,15 +8,14 @@ extension Generator {
     public func map<R>(_ transform: @escaping (T) -> R) -> AnyGenerator<R> {
         Map(source: self, transform: transform).eraseToAnyGenerator()
     }
-}
 
-// MARK: - Map
+    struct Map<Source, T>: Generator where Source: Generator {
+        let source: Source
+        let transform: (Source.T) -> T
 
-private struct Map<Source, T>: Generator where Source: Generator {
-    let source: Source
-    let transform: (Source.T) -> T
-
-    public func next(_ randomSource: RandomSource) -> T {
-        transform(source.next(randomSource))
+        public func next(_ randomSource: RandomSource) -> T {
+            transform(source.next(randomSource))
+        }
     }
+
 }
