+++ Sources/Genything/Operators/Generator+scan.swift
@@ -10,34 +10,32 @@ extension Generator {
         Scan(source: self, initialResult: initialResult, nextPartialResult: nextPartialResult)
             .eraseToAnyGenerator()
     }
-}
 
-// MARK: - Scan
+    final class Scan<Source, R>: Generator where Source: Generator {
 
-private final class Scan<Source, R>: Generator where Source: Generator {
+        init(
+            source: Source,
+            initialResult: R,
+            nextPartialResult: @escaping (R, Source.T) -> R
+        ) {
+            self.source = source
+            self.initialResult = initialResult
+            self.nextPartialResult = nextPartialResult
+        }
 
-    init(
-        source: Source,
-        initialResult: R,
-        nextPartialResult: @escaping (R, Source.T) -> R
-    ) {
-        self.source = source
-        self.initialResult = initialResult
-        self.nextPartialResult = nextPartialResult
+        public func next(_ randomSource: RandomSource) -> R {
+            let nextResult = nextPartialResult(previousResult, source.next(randomSource))
+            previousResult = nextResult
+            return nextResult
+        }
+
+        typealias T = R
+
+        let source: Source
+        let initialResult: R
+        let nextPartialResult: (R, Source.T) -> R
+
+        lazy var previousResult = initialResult
+
     }
-
-    public func next(_ randomSource: RandomSource) -> R {
-        let nextResult = nextPartialResult(previousResult, source.next(randomSource))
-        previousResult = nextResult
-        return nextResult
-    }
-
-    typealias T = R
-
-    let source: Source
-    let initialResult: R
-    let nextPartialResult: (R, Source.T) -> R
-
-    lazy var previousResult = initialResult
-
 }
