+++ Sources/Genything/Operators/Produce/Generator+sequence.swift
