+++ Sources/Genything/Operators/Produce/Generator+take.swift
